refactor: moving ready methods to NetworkSceneManager
BREAKING CHANGE: moving SetAllClientsNotReady and SetClientNotReady from ServerObjectManager to NetworkSceneManager
