+++ Assets/Mirage/Runtime/ServerObjectManager.cs
@@ -59,7 +59,6 @@ namespace Mirage
 
                 if (NetworkSceneManager != null)
                 {
-                    NetworkSceneManager.OnServerStartedSceneChange.AddListener(OnStartedSceneChange);
                     NetworkSceneManager.OnServerFinishedSceneChange.AddListener(OnFinishedSceneChange);
                 }
             }
@@ -89,11 +88,6 @@ namespace Mirage
             Server.World.ClearSpawnedObjects();
             // reset so ids stay small in each session
             nextNetworkId = 1;
-        }
-
-        void OnStartedSceneChange(string scenePath, SceneOperation sceneOperation)
-        {
-            SetAllClientsNotReady();
         }
 
         void OnFinishedSceneChange(string scenePath, SceneOperation sceneOperation)
@@ -674,34 +668,6 @@ namespace Mirage
                 SpawnObserversForConnection(player);
         }
 
-        /// <summary>
-        /// Marks all connected clients as no longer ready.
-        /// <para>All clients will no longer be sent state synchronization updates. The player's clients can call ClientManager.Ready() again to re-enter the ready state. This is useful when switching scenes.</para>
-        /// </summary>
-        public void SetAllClientsNotReady()
-        {
-            foreach (INetworkPlayer player in Server.Players)
-            {
-                SetClientNotReady(player);
-            }
-        }
-
-        /// <summary>
-        /// Sets the client of the connection to be not-ready.
-        /// <para>Clients that are not ready do not receive spawned objects or state synchronization updates. They client can be made ready again by calling SetClientReady().</para>
-        /// </summary>
-        /// <param name="player">The connection of the client to make not ready.</param>
-        public void SetClientNotReady(INetworkPlayer player)
-        {
-            if (player.SceneIsReady)
-            {
-                if (logger.LogEnabled()) logger.Log("PlayerNotReady " + player);
-                player.SceneIsReady = false;
-                player.RemoveAllVisibleObjects();
-
-                player.Send(new SceneNotReadyMessage());
-            }
-        }
 
         /// <summary>
         /// default ready handler.
