+++ Assets/Tests/Runtime/Host/PlayerSceneReadyHostTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d2a1a8dde706634fa570430b7dc0ff0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
